Move experiment components into experiments folder
