+++ client/src/components/experiments/IntervalChart.jsx
